+++ app/src/main/java/com/rakutentest/android/data/repository/dataSource/product/ProductRemoteDataSource.kt
@@ -11,6 +11,6 @@ interface ProductRemoteDataSource {
     ): Response<ProductList>
 
     suspend fun getProductDetails(
-        id: Int
+        id: Long
     ): Response<ProductDetails>
 }
