+++ app/src/main/java/com/rakutentest/android/domain/useCase/product/GetRemoteProductDetailsUseCase.kt
@@ -5,7 +5,7 @@ import com.rakutentest.android.data.util.Resource
 import com.rakutentest.android.domain.repository.ProductRepository
 
 class GetRemoteProductDetailsUseCase(private val productRepository: ProductRepository) {
-    suspend fun execute(id: Int): Resource<ProductDetails> {
+    suspend fun execute(id: Long): Resource<ProductDetails> {
         return productRepository.getProductDetails(id = id)
     }
 }
